requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.1.3
colorama==0.4.4
dnspython==2.2.1
fastapi==0.75.2
h11==0.13.0
idna==3.3
passlib==1.7.4
pydantic==1.9.0
PyJWT @ file:///home/conda/feedstock_root/build_artifacts/pyjwt_1652398519695/work
pymongo==4.1.1
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.2.0
uvicorn==0.17.6
